完成了人脸认证的通信和相关的页面跳转逻辑,完成了个人和修改界面的头像实时更新,增加开屏广告和广告相关的网络层，优化下拉刷新代码，

+++ app/src/main/java/com/example/welcome_freshman/feature/main/profile/Nav.kt
@@ -16,7 +16,7 @@ fun NavController.navigateToProfile(navOptions: NavOptions? = null) {
     this.navigate(PROFILE_ROUTE, navOptions)
 }
 
-fun NavGraphBuilder.profileScreen(onAuthenticationClick: () -> Unit,onUpdateUserClick: (Int) -> Unit) {
+fun NavGraphBuilder.profileScreen(onAuthenticationClick: () -> Unit,onUpdateUserClick: () -> Unit) {
     composable(
         route = PROFILE_ROUTE
     ) {
